增加系管理、班级管理、课程管理、健康管理的异常处理，并在前台页面上使用Alert将其输出。

+++ src/main/java/com/houseyoung/student/service/impl/CourseServiceImpl.java
@@ -18,32 +18,70 @@ public class CourseServiceImpl implements CourseService {
     private CourseMapper courseMapper;
 
     @Override
-    public List<Course> listCourse(String keywords){
-        return courseMapper.listCourse(keywords);
+    public List<Course> listCourse(String keywords) throws Exception{
+        try {
+            return courseMapper.listCourse(keywords);
+        } catch (Exception e){
+            throw new Exception("系统内部异常");
+        }
     }
 
     @Override
-    public void insert(Course course){
-        courseMapper.insert(course);
+    public void insert(Course course) throws Exception{
+        try {
+            if (course.getName() == null || course.getName() == "") {
+                throw new Exception("课程名不可为空");
+            }
+            //TODO: 判断是否重名
+            courseMapper.insert(course);
+        } catch (Exception e){
+            throw e;
+        }
     }
 
     @Override
-    public void delete(int id)  {
-        courseMapper.delete(id);
+    public void delete(int id) throws Exception{
+        try {
+            if (id <= 0) {
+                throw new Exception("ID不合法");
+            }
+            courseMapper.delete(id);
+        } catch (Exception e){
+            throw e;
+        }
     }
 
     @Override
-    public Course queryById(int id) {
-        return courseMapper.queryById(id);
+    public Course queryById(int id) throws Exception{
+        try {
+            if (id <= 0) {
+                throw new Exception("ID不合法");
+            }
+            return courseMapper.queryById(id);
+        } catch (Exception e){
+            throw e;
+        }
     }
 
     @Override
-    public void update(Course course){
-        courseMapper.update(course);
+    public void update(Course course) throws Exception{
+        try {
+            if (course.getName() == null || course.getName() == "") {
+                throw new Exception("课程名不可为空");
+            }
+            //TODO: 判断是否重名
+            courseMapper.update(course);
+        } catch (Exception e){
+            throw e;
+        }
     }
 
     @Override
-    public List<Course> listHimself(String studentId) {
-        return courseMapper.listHimself(studentId);
+    public List<Course> listHimself(String studentId) throws Exception{
+        try {
+            return courseMapper.listHimself(studentId);
+        } catch (Exception e){
+            throw new Exception("系统内部异常");
+        }
     }
 }
