+++ src/main/java/com/houseyoung/student/service/impl/HealthServiceImpl.java
@@ -18,32 +18,70 @@ public class HealthServiceImpl implements HealthService {
     private HealthMapper healthMapper;
 
     @Override
-    public List<HealthDto> listHealth(int classId, String keywords){
-        return healthMapper.listHealth(classId, keywords);
+    public List<HealthDto> listHealth(int classId, String keywords) throws Exception{
+        try {
+            return healthMapper.listHealth(classId, keywords);
+        } catch (Exception e){
+            throw new Exception("系统内部异常");
+        }
     }
 
     @Override
-    public void insert(HealthDto healthDto){
-        healthMapper.insert(healthDto);
+    public void insert(HealthDto healthDto) throws Exception{
+        try {
+            if (healthDto.getStudentId() == null || healthDto.getStudentId() == "") {
+                throw new Exception("学号不可为空");
+            }
+            //TODO: 判断学号是否存在
+            healthMapper.insert(healthDto);
+        } catch (Exception e){
+            throw e;
+        }
     }
 
     @Override
-    public void delete(int id)  {
-        healthMapper.delete(id);
+    public void delete(int id) throws Exception{
+        try {
+            if (id <= 0) {
+                throw new Exception("ID不合法");
+            }
+            healthMapper.delete(id);
+        } catch (Exception e){
+            throw e;
+        }
     }
 
     @Override
-    public HealthDto queryById(int id) {
-        return healthMapper.queryById(id);
+    public HealthDto queryById(int id) throws Exception{
+        try {
+            if (id <= 0) {
+                throw new Exception("ID不合法");
+            }
+            return healthMapper.queryById(id);
+        } catch (Exception e){
+            throw e;
+        }
     }
 
     @Override
-    public void update(HealthDto healthDto){
-        healthMapper.update(healthDto);
+    public void update(HealthDto healthDto) throws Exception{
+        try {
+            if (healthDto.getStudentId() == null || healthDto.getStudentId() == "") {
+                throw new Exception("学号不可为空");
+            }
+            //TODO: 判断学号是否存在
+            healthMapper.update(healthDto);
+        } catch (Exception e){
+            throw e;
+        }
     }
 
     @Override
-    public List<HealthDto> listHimself(String studentId) {
-        return healthMapper.listHimself(studentId);
+    public List<HealthDto> listHimself(String studentId) throws Exception{
+        try {
+            return healthMapper.listHimself(studentId);
+        } catch (Exception e){
+            throw new Exception("系统内部异常");
+        }
     }
 }
